Use Latte templates for visible pages
Use [Latte][1] templates for the login page and overview page. The auth
endpoint does not use templates for performance reasons - its contents
are irrelevant for Traefik, which only checks the status code.

Re-display the login page with an error message when the credentials
were wrong.

Use the [Pure CSS][1] framework for form formatting and general styling

[1]: https://latte.nette.org/
[2]: https://purecss.io/